backend/requirements.txt
fastapi
uvicorn
pydantic
reportlab
python-multipart
google-generativeai
python-dotenv
PyPDF2
supabase
